Remove exhausted retries logic
Recovering individual emails will prevent emails from being stuck in a
perpetually `pending` state and so the sidekiq exhausted retries block
can be removed which originally prevented this from happening. All
emails get deleted after 7 days so they have a maximum of 7 days to send
before getting deleted from the system entirely.

Another reason to remove this logic is that it would conflict with our
new alerting: we propose to alert on 5+ retries on pending emails, so if
this was left in then the alert would incorrectly fix itself once the
retries get exhausted.