Revert "Move rate_limiter out of Services module"
This reverts commit 8143521851bbe5de4eb2fe2ccd3b80f8d99693db.

We seem to have picked up an intermittent performance issue due to this
change that we've not been able to diagnose.

We have seen the occasional SendEmailWorker take crazy amounts of time
(around 4 hours). In trying to diagnose this on integration we found it
relatively easy to have a pool of workers take > 10 mins to
SendEmailWorker. We found that this 10 minutes was spent incrementing
the RateLimiter. Reverting this commit resolved this problem.

I'm not sure where the time gets lost in this, certainly not for hours.
I'm assuming there is some sort of deadlock that occurs which is
resolved by a timeout. This look a total pain to debug and we don't
have much time left as a team so I'm taking the rather ignorant route of
reverting without understanding exactly why the problem occurs.