Remove code and tasks for load testing
There aren't currently plans for future work on this application and
thus not a foreseeable need for load testing. This code is therefore
being removed from the repo.

This code will be linked to from our documentation so that this can be
resurrected should the need to load test return.