Adds global max Sidekiq retry count 4 all workers
Adds [global Sidekiq retry count] so our approach to native Sidekiq
retrying is consistent throughout the app. This also prevents most of
the workers retying for weeks due to the default exponential retry back
off (20 days) which we found excessive.

We already retry 9 times in the SendEmailWorker so this seemed a
reasonable global default to start at (total retry duration of 2h 39m
42s).

Trello:
https://trello.com/c/ssSnzp2L/640-remove-alert-from-all-the-sign-up-links

[global Sidekiq retry count]: https://github.com/mperham/sidekiq/wiki/Error-Handling#automatic-job-retry