Remove custom retry logic from initiators
Removed the custom retry count to use the global retry count (9). As
both of these workers run with locks we're protected against concurrent
running between the worker's native retries and the retries from within
the RecoverLostJobsWorker.

This custom retry logic is slightly odd as the initiator workers are the
only workers to use this approach so the relating code has been removed.