Remove link to feedback survey in emails
This is being removed since the Notifications team is being paused,
and there is no one to look at the feedback collected. While it may
normally be useful to continuously gather feedback, we also know that
the questions in the current survey are flawed i.e. there's no point
someone filling it in, because the data isn't useful.

Note that, if a user unsubscribes after an individual-type email is
generated, they will see a pointless separator at the bottom of the
email, due to the way the builder for this email is structure. This
doesn't like a big issue, and we're intending to rewrite the template
for them anyway, so I've left it that way for now.