Stop logging every email (fills up the disk!)
Previously we emulated the behaviour of the former PseudoProvider
by writing pseudo-sent email contents to a log file. However, this
never used to happen in Integration/Staging, since we actually used
the NotifyProvider with an allowed list [1]. This removes some of
the logging, since the sheer volume of logs has actually filled
one of the disks in Integration.

[1]: https://github.com/alphagov/email-alert-api/pull/1489