Reinstate confirmation email for subscriptions
Previously we removed this because it adds no value for a standard
GOV.UK subscription. However, we want it back for subscriptions to
Brexit Checker emails since the list description included in the
email contains a link back to the results page.

Note this makes a few changes to the original functionality:

- We now send the email if a previous subscription was overridden,
since the user clearly has no awareness of this.

- We now have the ability to skip sending the email, to prevent a
duplicate for users who signed up with a GOV.UK Account.