DRY-up creating/updating a subscription
Previously we had similar and voluminous logic for this in both the
"create" and "update" controller actions. This centralises the logic
into a new service, with a few functional changes:

- We now query an existing subscription twice, when coming via the
"update" action. This is actually beneficial: we now perform the
update on the data retrieved *after* locking the subscriber.

- Both actions benefit from error handling if the specified frequency
is invalid, as well as other invalid parameters when coming via the
"create" action.

- The Subscriber is now created outside of the transaction, as part
of the "create" action. This is OK because it's valid for the subscriber
to exist irrespective of the subscription.

- We now always return a "200" response for the "create" action, due
to the limited interface of the new service. This seems OK, since we
aren't using the response code for anything.