Assume we have subscriptions when building emails
Previously the builders for immediate-type emails had to cope with
the case of no subscriptions, in order to build the courtesy copy
emails. This removes the code to handle that case.

Since the builder specs were previously written assuming one could
pass no subscriptions, I had to rewrite them so that we're always
passing subscriptions when we call the builder.

Note that, while it's possible for someone to unsubscribe as we're
generating email for them, we already filter out such subscribers
in the calling code [1] [2].

[1]: https://github.com/alphagov/email-alert-api/blob/0e9b38092921a81e729ad317ee2bd8e62e2252ff/app/services/immediate_email_generation_service/batch.rb#L63
[2]: https://github.com/alphagov/email-alert-api/blob/0e9b38092921a81e729ad317ee2bd8e62e2252ff/spec/services/immediate_email_generation_service/batch_spec.rb#L104