Update confirmation email to match new design
Note that this generalises the locale key used for the footer, since
we should try to keep it consistent across the emails we send.

Note also, as in 0264407b, we no longer make the subscriber list title
a link, as this is inconsistent and incompatible with the new design
for digest emails, where list title are used as a contents list.