Switch to one digest email per subscription
Previously we aggregated all subscriptions of the same frequency
into the same digest email. This split them into separate emails
to work towards the new design:

- Having one group of content per email is simpler.
- We can have a single, one-click unsubscribe link in the footer.

Note that, at the time of writing, only ~89% of digest subscribers
have only one digest subscription, so this change is to optimise
the digest experience for the majority of subscribers.

Note also that this will mean some workers are occupied creating
multiple emails, which we currently do with individual INSERT
statements. This seems reasonable for now, as it's bounded by the
number of subscriptions for the subscriber. We could always change
it to do a batch insertion [1] if needed.

[1]: https://github.com/alphagov/email-alert-api/blob/1cd8a6e50782f1ae2932a31a2cf194bebdceafe3/app/workers/digest_email_generation_worker.rb#L59