Extend individual email recovery time
We will be removing [Sidekiq unique jobs] from the SendEmailWorker as
we suspect it of causing a high load (and slowdown) on Redis when there is
a large number of emails queued up to be sent.

This change extends the time before we recover individual emails from 1
hour to 3 hours. We're extending this time to give the email system a
bit of breathing room before bombarding it with recovery emails as
without a uniqueness constraint on the email job this leaves the
possibility of, during a busy (or unhealthy) period, duplicate emails
being requeued over and over again potentially causing huge queues and
further issues.

[Sidekiq unique jobs]: https://github.com/mhenrixon/sidekiq-unique-jobs