Remove Sidekiq uniqueness from SendEmailWorker
This removes the uniqueness constraint on emails from the
SendEmailWorker. We have evidence that during busy periods using a
the uniqueness constraint substantially slows down the sending of
individual emails.

In a load test where approximately 6 million emails were queued up the
time taken for each job in the SendEmailWorker averaged 8/9 seconds
(wow) - our normal processing time is around 200ms so this is a large
loss of performance.

We tested removing the constraint for the next load test as we had
suspicions that it was the culprit and the SendEmailWorker's performance
returned to operated at normal levels.