Enhance notify helper to expect specific emails
Previously it was only possible to retrieve the latest email sent
to Notify using the helper, with some specs writing their own stubs
in order to match a specific one. This enhances the helper to match
by subject and address (an important aspect of all emails).

Doing this has the immediate benefit of DRYing-up existing specs,
and making them more consistent. We will also use it in the next
commit to simplify the specs for digest emails.