Define 'send_email_worker' alongside 'worker'
For now, this uses the same properties as those defined in sidekiq.yml,
as it seems sensible to stick to these tried and tested values. All that's
missing is the actual queues and cron instructions, so currently this
worker wouldn't actually do anything.

For local development, we'd need to set an additional worker in govuk-docker:
https://github.com/alphagov/govuk-docker/blob/da856818fb571c3c1a2e392603f5e24d143a12ed/projects%2Femail-alert-api%2Fdocker-compose.yml#L45-L53

For this to work on production, we'd also need to define the worker in govuk-puppet:
https://github.com/alphagov/govuk-puppet/blob/574a04d1058f9c623ee6f3dd696655969a96cfc9/modules/govuk/manifests/apps/email_alert_api.pp#L146-L152

It would largely be a copy and paste, but renaming 'email-alert-api' to
something like 'email-alert-api-send-email', which becomes the `$title`
used throughout the `govuk::procfile::worker` class (and also lends
itself to the service name we use in commands like
`service {SERVICE_NAME} start`):
https://github.com/alphagov/govuk-puppet/blob/fbbecba00f5e4a06bed1c48b34445d7e4235c0fd/modules%2Fgovuk%2Fmanifests%2Fprocfile%2Fworker.pp

We'd need to pass `process_type => 'send_email_worker'` so that it knows
which worker to instantiate, and also potentially a `process_regex` so
that collectd & Icinga work correctly. There's prior art in govuk-puppet
([1]) for Content Data API, whose process is called 'default-worker'
rather than 'worker' ([2]). We also used to have multiple workers in
the (now retired) Content Audit Tool ([3] and [4]), for which I remember
having to fix the regex on my first 2nd line shift ([5]).

[1]: https://github.com/alphagov/govuk-puppet/blob/7e6a90da615e1094df305a367418139a76de0f8d/modules%2Fgovuk%2Fmanifests%2Fapps%2Fcontent_data_api.pp#L147-L152
[2]: https://github.com/alphagov/content-data-api/blob/2df29c27f2e715b07f5b8758519bb0d01ebdf6ff/Procfile#L1
[3]: https://github.com/alphagov/content-audit-tool/commit/6edc4e488dd8d7347517f5818ad977d9755fba01#diff-0a99231995da379e7aebabe76c9d849a23737a42c3b3a8994043e2aa80958424
[4]: https://github.com/alphagov/govuk-puppet/blob/e335930e21814e7aa27efa7e8f4e5360796091e9/modules/govuk/manifests/apps/content_audit_tool.pp#L96-L106
[5]: https://github.com/alphagov/content-audit-tool/pull/306