Reduce queue priorities to their common factor
The previous values were relative to '1' (the 'default' queue in
sidekiq.yml). Now that no queues in sidekiq_send_email.yml have a
priority of 1, we should simplify to the common factor of the
fraction, so 2 becomes 1 and 8 becomes 4.

As before, for every 1 time Sidekiq checks the 'send_email_digest'
queue, it will check the 'send_email_immediate_high' queue 4 times.