Set concurrency values according to the relative priorities
In moving from 1 parent worker to 2, and copying the same configs,
I've doubled the number of workers we would have running, which
would almost certainly cause a load issue.

At time of writing, there are 12 email machines in production, each
running 32 workers. This is 384 workers in total.

These workers are currently responsible for both processing content
changes _and_ then sending the email. Prior to #1534, the relative
priorities for each task was set out in sidekiq.yml:

```
  - [send_email_transactional, 8] # To send one-off emails, such as subscription confirmation.
  - [send_email_immediate_high, 8] # To send high priority emails to users with immediate subscriptions e.g. travel advice.
  - [process_and_generate_emails, 4] # To generate emails to users with immediate subscriptions for content changes and messages.
  - [email_generation_digest, 4] # To generate emails to users with daily or weekly subscriptions for content changes and messages.
  - [send_email_immediate, 2] # To send emails to users with immediate subscriptions.
  - [send_email_digest, 2] # To send digest emails to users with either daily or weekly subscriptions.
  - [default, 1] # Miscellaneous operations e.g. initiating digest runs, monitoring, DB cleanup and recovering lost Sidekiq jobs.
```

If we think of these priorities as worker counts, we'd have 29
workers. Hypothetically, 8 of these workers would be processing
items in the 'send_email_transactional' queue, and 1 worker would
be processing items in the 'default' queue.

In this hypothetical example, 20 of the 29 workers are responsible
for sending emails. So in terms of concurrency values, we should
have a value of 22 in `sidekiq_send_email.yml` (because 20/29 * 32 = 22),
and 10 in `sidekiq.yml`. We'll still have 384 workers in total
across all machines, and they'll all be doing roughly the same
prioritised work as before.

We can tweak these values in response to load testing.