Resturcture bulk email builder to work in batches
Previously the builder held all data in memory, irrespective of the
number of emails being generated. This rewrites the builder to be
more memory-efficient, by working in terms of batches of subscribers,
much like the ImmediateEmailGenerationService [1].

Note that the query is now on the Subscription model, despite only
needing subscriber data. This will change in a later commit, as we
start to use subscription data as well. For now, it's a stepping stone
to make that later change easier to understand.

[1]: https://github.com/alphagov/email-alert-api/blob/e6b34a68e734f322d883c978e810185fbd8dbce9/app/services/immediate_email_generation_service.rb#L27-L37