Inject list URL into bulk emails
This adds a new feature to bulk emails where the specified body
contains a "%LISTURL%" placeholder, which is substituted with the
URL of the latest subscription / subscriber list, if one is defined,
and otherwise rendered as an empty link.

Note that we expect any usage of bulk email functionality to first
check whether a "%LISTURL%" placeholder makes sense, for the email
they are sending i.e. all the lists have a URL to inject. Otherwise,
the rendered link will be a link to nowhere.