Add tracking to substituted bulk email URL
The tracking attributes are based on those for content changes [1],
but with a unique campaign (since this is no ordinary email). Using
the list slug for the utm_source makes sense here, since the email
is essentially about the subscription itself.

Note that this changes the arbitrary behaviour if one of the lists
does not have a URL, due to the need to parse it in order to avoid
overriding any existing query params. If one of the lists does not
have a URL, we will now substitute an empty string.

[1]: https://github.com/alphagov/email-alert-api/blob/d80bdd48198ffa37dfabe5ba630daa6c63190676/app/presenters/content_change_presenter.rb#L26-L33