Add tracking for source URLs in all emails
This adds UTM tracking params to a source URL in the three emails
it may occur in. Using the subscriber list slug as the UTM source
is consistent with the existing tracking of source URLs in bulk
emails [1]. Including a UTM content param will differentiate the
three normal emails in which the source URL appears.

Note that future tracking of other links in digest emails, such as
"Unsubscribe", can also make use of the newly available subscriber
list slug in the digest item.

[1]: https://github.com/alphagov/email-alert-api/blob/e23d929716416fc0147334af486023d57370558d/app/presenters/bulk_email_body_presenter.rb#L17