Make subscription available in digest builder
This is so that we can DRY-up generating the email footer, together
with the other builders. In order to do this, we need to make the
way the footer is generated consistent across builders. Previously
the builder only had access to specific attributes of the subscriber
list and the subscription. This changes the way "digest items" are
generated, doing one extra query to get real subscription objects,
including the subscriber / list.

Note that here we also stop passing the DigestRun to the builder,
since it makes more sense to work with the subscription frequency,
as opposed to the range of the digest run, which was the only use
for the object in the builder. As a consequence, the worker spec
needs to specify the frequency of the subscriptions, which would
otherwise default to "immediately".