DRY-up email footers into a new presenter
Note that this changes the explanation text for bulk emails, which
was previously hard-coded to that for individual-type emails, which
was inaccurate for digest subscriptions. While this can still be
perceived as inaccurate - the bulk email is in addition to a digest
email - hopefully the content of the bulk email will explain itself.