Remove redundant "order" clause in digest query
Previously we arbitrarily ordered the results in the digest items
queries, despite this being later overridden by a call to "sort_by".
While the ordering does affect which content change is favoured in
the subsequent uniquification, it was still arbitrary. This removes
the arbitrary ordering in the query, since it was confusing to see
this in tandem with a subsequent call to "sort_by".