Update the comment for the "message is too long" fix
This comment has been updated to reflect the non-ideal nature of this
fix. It is presented with the hope it doesn't encourage people to add
similar things in future.

As I reference "stop-gap" in the commit message I have no doubt that I have
doomed this code to live in the code for another decade.