Fix incorrect UTM content param for content change
Previously the link to a content item in an individual-type email
would have a UTM content param of "immediate", which is inconsistent
with the actual name of the frequency ("immediately"), which we use
elsewhere. Although we could just amend the default argument of the
content change presenter, this goes further by passing the whole
subscription object to both presenters, which reduces the coupling
between them and the calling code.