Fix failing test for confirmation token
Follow-up to: https://github.com/alphagov/email-alert-api/pull/1559

Previously the test was passing erroneously [1]. While the code to
generate authentication tokens had changed [2], the final URL still
needed to be encoded, so removing "CGI.unescape" call entirely was
incorrect. This reverts the change at the integration level, where
we expect the URLs to be encoded correctly. Note that the decoding
method is different, to match the encoding in PublicUrls [3].

This also updates the similar test for token links in emails to
manage subscriptions. Such emails appear to have been unaffected,
due to the links being rendered in full, which causes the browser
to automatically encode them before the user clicks on them.

[1]: https://github.com/alphagov/email-alert-api/pull/1548/commits/57f2fca2d83391b88c75aaa3b9862b470715df00
[2]: https://github.com/alphagov/email-alert-api/pull/1548/commits/57f2fca2d83391b88c75aaa3b9862b470715df00#diff-8c42a5fcffdef2c573e42bded73714a50ccc341c31701c6d68a2272029a4279dL18
[3]: https://github.com/alphagov/email-alert-api/commit/c2c5a3fb41169a18db1cf7aae5f67d7343fb2005#diff-0c46a2c123afab8047a264fefb57b5eec7186e55c3613b11351f46713eae7419R8