Simplify de-dup in ImmediateEmailGenerationService
Previously the service used "subscription_ids_by_subscriber" and
had some quite complex code to map slices of this hash into actual
model instances, for use in the builder. This replaces much of the
complexity with the knowledge that the builder only / arbitrarily
uses the first subscription for each subscriber [1].

Note that an subtle effect of this is that we now only create a
SubscriptionContent record for one subscription, even if multiple
subscriptions are relevant for a given subscriber and the content
change / message. This is OK because the SubscriptionContent record
is only used to associate generated emails with subscribers so we
know which subscribers have been "fulfilled".

[1]: https://github.com/alphagov/email-alert-api/blob/5a4dc5826ee5d35fb5e2b33b3f304ccf6a84961e/app/builders/immediate_email_builder.rb#L26