DRY-up adapting params into matching criteria
Previously we copied "find_exact_params" into the service from the
controller, with the aim of removing it there, along with the "index"
endpoint. However, the endpoint is still needed, so this modifies
the interface of the service to accept the params from the controller.

Note that the test for coping with legacy params is deleted in this
commit. We will re-add it at the integration level in the following
commits, after a bit of refactoring to make that easier.