Extend feature spec for subscribing to a list
Previously this only covered the latter part of the journey. This
rewrites the spec to represent that two API endpoints are used in
practice: one to send a verification email, and the other to create
the subscription once the email has been verified.

Note that this also separates the previous additional expectation -
around creating a duplicate subscription - into a separate scenario.
The other expectation, relating to a 404, is not appropriate at the
feature level - it's still tested at the integration level [1]

[1]: https://github.com/alphagov/email-alert-api/blob/4075aba2562f667bdd132a57453225e1116ab349/spec/integration/create_subscription_spec.rb#L32