Convert to date in SubscriberListsReport
A time-at-midnight-with-a-timezone compares as less than a date:

    irb(main):001:0> with_ts = Time.zone.parse(Time.zone.today.to_s)
    => Mon, 29 Mar 2021 00:00:00.000000000 BST +01:00

    irb(main):002:0> without_ts = Time.zone.today
    => Mon, 29 Mar 2021

    irb(main):003:0> with_ts < without_ts
    => true

So the check that the date must be in the past isn't working right: a
test that `Time.zone.today` gets rejected is failing, because the time
is being converted through `to_s`/`Time.zone.parse`, and ends up as
comparing less than its original value.

This report uses the date, not the time, so convert the input back
into a timezone-less date object.

Using `Time.zone.today.in_time_zone` for the comparison instead also
works.  I don't think there's a difference in practice, because the
`@date` gets used in database queries in the context of
`beginning_of_day` and `end_of_day`, which are timezone aware:

    irb(main):001:0> Time.zone.today.beginning_of_day
    => Mon, 29 Mar 2021 00:00:00.000000000 BST +01:00
    irb(main):002:0> Time.zone.today.in_time_zone.beginning_of_day
    => Mon, 29 Mar 2021 00:00:00.000000000 BST +01:00

    irb(main):003:0> Time.zone.today.end_of_day
    => Mon, 29 Mar 2021 23:59:59.999999999 BST +01:00
    irb(main):004:0> Time.zone.today.in_time_zone.end_of_day
    => Mon, 29 Mar 2021 23:59:59.999999999 BST +01:00