Use a mixin for common `.call` class interface
This creates a new mixin, Callable, which can be used to provide a class
with the boilerplate so that it can have a public class method of
`.call` which calls `#initialize` with the arguments and then `#call`.

The creation of this class is a port of the same approach applied in
Content Publisher [1] (and most of this commit message is a port too in
case you're feeling a touch of déjà vu).

This mixin replaces the inheritance approach used in builders,
presenters and services.

The change was prompted by a linting violation with one of rubocop's newer
rules: `Lint/MissingSuper`, which is an offence when a class inherits but
does not call `super` in it's initialize method. This violation seemed
to only affect some of GOV.UK's newer code, such as this service
pattern, which led to some debate [2] about what the most
appropriate means to share this behaviour is, whether our
inheritance approach was actually idiomatic and how using super in every
`initialize` would be unnecessary boilerplate.

The Rubocop thread [3] regarding this rule raised an interesting point
"I'm curious of examples where there are good reasons to explicitly not
call super.", which led to me considering why we didn't think `super`
would be appropriate in these classes. My conclusion to this was that we
weren't actually intending to create a subtype with inheritance, we
actually just wanted standardisation with the interface, and that the
inherited classes have very little behavioural consistency, breaking
Liskov substitution principle [4]. When looking at the Ruby standard
library a module seems the most appropriate way to implement it,
considering similarities with Singleton and Enumerable.

[1]: https://github.com/alphagov/content-publisher/commit/5933f87c02403e25e06af6152b9f348d3235b2b2
[2]: https://github.com/alphagov/rubocop-govuk/pull/125
[3]: https://github.com/rubocop/rubocop/issues/8506
[4]: https://thoughtbot.com/blog/back-to-basics-solid#liskov-substitution-principle