Fix Lint/ConstantDefinitionInBlock violations
These violations occur because a class defined with a block scope is
available in the same scope as the block. In these tests cases as there
is no scope above the block one these classes are defined globally.

To fix these an anonymous class is created as part of a let.

While fixing this I also went and made these tests more conventional for
RSpec by dropping the shorthand `is_expected`. I also went through and
removed the tests that I felt were superfluous where they tested logic
that wasn't in the code under test or tested the same thing multiple
times.

I do have some concerns about the validity of the TagsValidator but
that's a yak I didn't want to shave today.