Provide some feedback to the bulk email send task
Prior to this commit, the only feedback we'd have as a result of
triggering the task was `Finished: SUCCESS`. It was impossible to
know whether, for example, no subscriber lists were matched and
therefore no emails sent, in which case the task would be failing
silently. Giving an email count gives some indication of whether
something happened.

Trello: https://trello.com/c/qqxUkxPX/2489-2-send-email-to-all-gds-department-subscribers-pointing-them-to-cddo-alerts