Import email-alert-api docs into email-alert-api/docs
Importing from the govuk-developer-docs so that the docs can live
closer to the code, making it easier to make changes to both at
the same time, as well as to have relevant email-alert-api docs
locally when checking out the repo. This was not possible prior to
https://github.com/alphagov/govuk-developer-docs/pull/2910, which
enabled external doc importing into the govuk-developer-docs.

Future devs can dig into the commit history of the govuk-developer-docs
versions of these files at these URLs:

- https://github.com/alphagov/govuk-developer-docs/blame/2289779c12930502177bc2d43c10dd385c834c90/source/manual/load-test-email-alert-api.html.md
- https://github.com/alphagov/govuk-developer-docs/blame/2289779c12930502177bc2d43c10dd385c834c90/source/manual/receiving-emails-from-email-alert-api-in-integration-and-staging.html.md
- https://github.com/alphagov/govuk-developer-docs/blame/2289779c12930502177bc2d43c10dd385c834c90/source/manual/bulk-email.html.md