Expand FindExactQuery to handle content_ids (#1640)
There is precedent here for other simple matches.

It seems we can already query by document_type so lets follow the logic
and the tests there.

We should return a subscriber list if content_id matches what is passed
to it as a query.

We should not return a subscriber list if content_id fails to match.

We also need additional tests to show the overlap with :links and :tags.
These are an AND relationship.

So:

if `content_id` AND `:tags` match, return subscriber list
if `content_id` AND `:links` match, return subscriber list

However if we cannot match on both of those, return nil - even if we
have a matching content ID.

If you are querying for both, you should only match if you have both.