When logging in with a GOV.UK account, find or create the subscriber
We've decided that the states "no subscriber exists" and "a subscriber
with no subscriptions exists" should be the same.

To remove the 404 status, I've changed the no-email case to return a
403.  This case will only happen if the account-api doesn't have that
attribute, which could only be the case for a brief period after the
user registers, in which case they won't be verified either.