Allow merging subscribers on address clash
As part of bringing notifications and accounts together, we need to be
able to automatically handle the case where a user:

1. Has notifications set up for `foo` and `bar`
2. Registers a GOV.UK account to manage their notifications for `foo`
3. Changes the email address of their account to `bar`

Currently, email-alert-api will reject this change: it doesn't let you
change to an email address that's already in use.

But for the change to become active, the user has proven to the
account that they own both `foo` (we're not letting them do any
notifications things without confirming their email address) *and*
`bar` (we won't update the address until it's been confirmed).

We've decided in this case to merge the two subscribers in
email-alert-api.  This is pretty straightforward, other than the case
where both subscribers have an active subscription to the same topic,
with different frequencies.  We've decided to keep the most frequent
if that happens.  The other is marked as ended with a new
`subscriber_merged` reason, so we can see how much this happens.

This new behaviour is only used if `on_conflict=merge` is passed along
as a parameter.  We'll make account-api do that, but
email-alert-frontend will retain the old behaviour.