Allow bulk email to skip arbitrary addresses
While we retire the Brexit checker, we need to do the following:

- send an email to subscribers with a GOV.UK account
- send a different email to subscribers without a GOV.UK account

The govuk_account_id property of a subscriber is no longer populated,
otherwise it would have been exactly the thing to use. Unfortunately, this
information is now held within the accounts-api.

We're going to extract a list of email addresses, and replace the example
file, so that we can send the email to non-holders of a GOV.UK account.

We'll use the same file to run the opposite task for the other set of users
which is coming in the next commit.