Allow bulk email to only send to specified addresses
This allows us to send a specific bulk email to members of a subscriber list
but only if their email address is in the file in config/bulk_email/email_addresses.txt.

This is to allow us to segment the email that informs users of the retirement
of the Brexit checker into those people with an account and those without.

The spec file is 99% a copy and paste of bulk_subscriber_list_email_builder_spec.rb
as the funtionality is basically identical apart from the inverse operation
of the `accept_required_subscribers` method. I tried it with shared examples,
but found it less readable.

Also, I'm going to revert these commits once this has been sent. We may need it
again, but I'd prefer a better solution to exist in future.