Add link to confirmation email for single page subscriptions
So that it's easy for users to get back to the page. Also rename the
factory trait added in #1679 - we are interested in the subscriber_list
being for a single page, not only that it has a content_id (which is a
side effect of being for a single page).