Destroy all Brexit checker subscriber lists
We deactivated all the subscriptions last week.

This now allows us to delete the subscriber lists, subscriptions and
subscription contents from the database. The subscriber lists in particular
are the bulk of the content in that table.

We could do a couple of delete statements to achieve the same sort of thing
more quickly but it's not important for this to be a fast script. (It'll
run individual delete statements for each row.) What's more important is
to avoid locking on the production database as this may cause disruption to
other activities. The other benefit of using `destroy` is that it removes
the related objects through `dependent: :destroy` annotations on the models
without us having to worry about it.