Add `omit_footer_unsubscribe_link` to Message model
We're re-using this model for emails which are sent when a
`SubscriberList` is bulk-unsubscribed (such as when we automatically
remove the list for a single-page subscription when the piece of
content gets unpublished).

Since the user is unsubscribed in that case, we don't want to include
the standard unsubscribe link in the footer as well.