Fix sending unsubscribe notifications immediately
We use the message machinery to send a notification to single-page
subscribers when the page gets unpublished and we unsubscribe them
from it.

We need these emails to go out immediately so we know when to
unsubscribe the user.  So in this commit, I add an
`override_subscription_frequency_to_immediate` field to the `messages`
table, and use that to skip the frequency scope check in the
`ImmediateEmailGenerationService`.

Alternative approaches would be:

- We could change the frequency of the `Subscription` when we process
the bulk-unsubscription.  This doesn't feel in keeping with the spirit
of email-alert-api, where normally data is not mutated in cases like
that.

- We could include unsubscription notifications in batches.  But then
we don't know when we can end the subscription (worst case we have to
wait a week), and if the page is gone anyway the other updates in the
batch are likely useless.

I think this is a least-bad approach.  But it is a feature we should
be careful not to abuse for other purposes.  We offer users frequency
controls for a reason.