Add Patch endpoint for subscriber lists
An endpoint to do a partial update for subscriber lists, specifically
targeting metadata we expose to the user to help them understand what
the list they are subscribed to targets.

When we added single page notiications we allowed subscriber lists to be
created for individual pages the first time a user subscribes.

A title is added at that point, which is then used in unpublication
emails, and rendered in the user's account to help them manage their
subscriptions.

Until this PR there wasn't a way to update these titles on subscriber
lists. This is a problem as individual pages do have titles that change
over time, and users may be confused if they see they are subscribed to
a page with a title they do not recognise.