Add rake task to add descriptions to subscriber lists
We added an optional description field to subscriber lists in [[1]]. We
intend to use the page summary as this description for single page
subscriber lists, and will include it in some emails.

New single page lists get the description passed through from email
alert frontend when the list is created, but it's null for any that
existed before this change.

Add a rake task that iterates over all the single page subscriber lists
and fetches the description from content store.

[1]: https://github.com/alphagov/email-alert-api/pull/1705